# Picking latest gas/token price - PoC