### Fetching latest gas price

```go
//config
numberOfLogs := 20_000
```

**Without limit 86 millis on avg and 20k logs fetched to memory**
```
Benchmark_SelectIndexedLogsCreatedAfter
Benchmark_SelectIndexedLogsCreatedAfter-12    	      14	  89969223 ns/op
Benchmark_SelectIndexedLogsCreatedAfter-12    	      13	  82716218 ns/op
Benchmark_SelectIndexedLogsCreatedAfter-12    	      14	  80500387 ns/op
Benchmark_SelectIndexedLogsCreatedAfter-12    	      14	  89134684 ns/op
Benchmark_SelectIndexedLogsCreatedAfter-12    	      12	  89081694 ns/op
```

**With limit applied 11 millis on avg and 1 log fetched to memory**
```
Benchmark_SelectIndexedLogsCreatedAfter
Benchmark_SelectIndexedLogsCreatedAfter-12    	     104	  11814839 ns/op
Benchmark_SelectIndexedLogsCreatedAfter-12    	      92	  11863810 ns/op
Benchmark_SelectIndexedLogsCreatedAfter-12    	      91	  12101903 ns/op
Benchmark_SelectIndexedLogsCreatedAfter-12    	      93	  12057668 ns/op
Benchmark_SelectIndexedLogsCreatedAfter-12    	     103	  11095903 ns/op
```

### Fetching multiple token prices

```go
//config
numberOfLogs := 20_000
numberOfTopics := 100 // 100 tokens
```
**LogsCreatedAfter 86 millis avg and 20k logs fetched to memory**
```
Benchmark_SelectLatestIndexedLogs
Benchmark_SelectLatestIndexedLogs-12    	      13	  85237070 ns/op
Benchmark_SelectLatestIndexedLogs-12    	      13	  90363760 ns/op
Benchmark_SelectLatestIndexedLogs-12    	      12	  90033184 ns/op
Benchmark_SelectLatestIndexedLogs-12    	      13	  83897032 ns/op
Benchmark_SelectLatestIndexedLogs-12    	      13	  81574080 ns/op

```

**SelectLatestIndexedLogs 46 millis avg and 100 logs fetched to memory**
```
Benchmark_SelectLatestIndexedLogs
Benchmark_SelectLatestIndexedLogs-12    	      27	  46940284 ns/op
Benchmark_SelectLatestIndexedLogs-12    	      26	  47423162 ns/op
Benchmark_SelectLatestIndexedLogs-12    	      24	  51321167 ns/op
Benchmark_SelectLatestIndexedLogs-12    	      24	  42984066 ns/op
Benchmark_SelectLatestIndexedLogs-12    	      24	  44422215 ns/op
```